PycharmProjects/rapid_eld/requirements.txt
# Core dependencies (required)
requests==2.31.0

# Optional dependencies for different integrations
# Flask integration
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0

# Celery for background tasks
Celery==5.3.4
redis==5.0.1

# Configuration
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-flask==1.2.0
